Dev stable, forgot to merge bunch of stuff

+++ src/main/java/com/example/um/Building/BuildingController.java
@@ -1,6 +1,8 @@
 package com.example.um.Building;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import com.example.um.Campus.CampusRepository;
+import com.example.um.User.User;
+import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -10,11 +12,14 @@ import java.util.stream.Collectors;
 
 
 @RestController
+@CrossOrigin(origins = "/**",  allowedHeaders = "Content-Type, Authorization")
 @RequestMapping("/api/buildings")
 public class BuildingController {
-    @Autowired
-    private BuildingService buildingService;
+    private final BuildingService buildingService;
 
+    public BuildingController(BuildingService buildingService) {
+        this.buildingService = buildingService;
+    }
 
     @GetMapping
     public List<BuildingDTO> getAllBuildings() {
@@ -28,35 +33,50 @@ public class BuildingController {
 
     @GetMapping("/{id}")
     public ResponseEntity<BuildingDTO> getBuildingById(@PathVariable Long id) {
-        return buildingService.findBuildingById(id).map(building -> new BuildingDTO(
-                building.getId(),
-                building.getCode(),
-                building.getYearOfConstruction(),
-                building.getCampus().getId())).map(ResponseEntity::ok).orElseGet(()->ResponseEntity.notFound().build());
+        return buildingService.findBuildingById(id)
+                .map(building -> new BuildingDTO(
+                        building.getId(),
+                        building.getCode(),
+                        building.getYearOfConstruction(),
+                        building.getCampus().getId()))
+                .map(ResponseEntity::ok)
+                .orElseGet(() -> ResponseEntity.notFound().build());
     }
 
     @PostMapping
-    public Building createBuilding(@RequestBody Building building) {
-        return buildingService.saveBuilding(building);
+    public ResponseEntity<Building> createBuilding(@RequestBody BuildingDTO building, @RequestHeader("Role") String role) {
+        User.Role enumRole = User.Role.valueOf(role);
+        if (!enumRole.equals(User.Role.ADMINISTRATOR) && !enumRole.equals(User.Role.MANAGER)) {
+            return ResponseEntity.status(403).body(null); // Forbidden
+        }
+        return ResponseEntity.status(200).body(buildingService.createBuilding(building));
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Building> updateBuilding(@PathVariable Long id, @RequestBody Building buildingDetails) {
-        Optional<Building> buildingOptional = buildingService.findBuildingById(id);
-        if (buildingOptional.isPresent()) {
-            Building building = buildingOptional.get();
-            building.setCode(buildingDetails.getCode());
-            building.setYearOfConstruction(buildingDetails.getYearOfConstruction());
-            building.setCampus(buildingDetails.getCampus());
-            return ResponseEntity.ok(buildingService.saveBuilding(building));
-        } else {
+    public ResponseEntity<Building> updateBuilding(@PathVariable Long id, @RequestBody BuildingDTO buildingDetails, @RequestHeader("Role") String role) {
+        try {
+            Building updatedBuilding = buildingService.updateBuilding(id, buildingDetails);
+            User.Role enumRole = User.Role.valueOf(role);
+            if (!enumRole.equals(User.Role.ADMINISTRATOR) && !enumRole.equals(User.Role.MANAGER)) {
+                return ResponseEntity.status(403).body(null); // Forbidden
+            }
+            return ResponseEntity.ok(updatedBuilding);
+        } catch (ResourceNotFoundException e) {
             return ResponseEntity.notFound().build();
         }
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteBuilding(@PathVariable Long id) {
-        buildingService.deleteBuilding(id);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<Void> deleteBuilding(@PathVariable Long id, @RequestHeader("Role") String role) {
+        try {
+            User.Role enumRole = User.Role.valueOf(role);
+            if (!enumRole.equals(User.Role.ADMINISTRATOR) && !enumRole.equals(User.Role.MANAGER)) {
+                return ResponseEntity.status(403).body(null); // Forbidden
+            }
+            buildingService.deleteBuilding(id);
+            return ResponseEntity.noContent().build();
+        } catch (ResourceNotFoundException e) {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
